requirements.txt
numpy==1.19.4
streamlit==0.73.1
pandas==1.1.5
matplotlib==3.3.3
Pillow==8.1.0
